Moved glfw_window.hpp to src/cubos
